+++ kernel/arch/x86/elf.cpp
@@ -48,8 +48,8 @@ void elf::load_program(void *ELF_baseadr, std::vector<std::string> *argv, bool r
 
     std::vector<multitasking::process_pagerange> pageranges;
 
-    pageranges.push_back({.phys_base = (uint32_t)mm::phys::phys_malloc(pagecount), .virt_base = min_v, .pages = pagecount, .type = multitasking::process_pagerange::range_type::STATIC});
-    pageranges.push_back({.phys_base = (uint32_t)mm::phys::phys_malloc(1), .virt_base = max_v + (ARCH_PAGE_SIZE * 47), .pages = 1, .type = multitasking::process_pagerange::range_type::BREAK});
+    pageranges.push_back({.phys_base = (uintptr_t)mm::phys::phys_malloc(pagecount), .virt_base = min_v, .pages = pagecount, .type = multitasking::process_pagerange::range_type::STATIC});
+    pageranges.push_back({.phys_base = (uintptr_t)mm::phys::phys_malloc(1), .virt_base = max_v + (ARCH_PAGE_SIZE * 47), .pages = 1, .type = multitasking::process_pagerange::range_type::BREAK});
 
     std::vector<multitasking::process_pagerange> old_pageranges;
     multitasking::createPageRange(&old_pageranges);
