+++ kernel/arch/x86/include/arch/drivers/net/rtl8139.h
@@ -3,6 +3,6 @@
 
 namespace drivers::net::rtl8139 {
     void init();
-    void sendPacket(void *data, uint32_t len);
+    void sendPacket(void *data, size_t len);
     uint8_t get_mac_byte(int n);
 }
